# backend/docs/swagger.yaml
basePath: /
definitions:
...
  domain.CreatePlaceRequest:
    properties:
      configuration_id:
        type: integer
      is_active:
        type: boolean
      label:
        type: string
      sort_order:
        type: integer
      zone_id:
        type: integer
    required:
    - label
    - zone_id
    type: object
...
  domain.PatchPlaceRequest:
    properties:
      configuration_id:
        type: integer
      is_active:
        type: boolean
      label:
        type: string
      sort_order:
        type: integer
    type: object
...
  response.DeletePlaceResponse:
    properties:
      id:
        type: integer
    type: object
...
  response.ErrorResponse:
    properties:
      code:
        example: auth_failed
        type: string
      details: {}
      message:
        example: Authentication failed
        type: string
      request_id:
        example: 7fbd6854-8e42-4451-80ee-6da60aeceacd
        type: string
      timestamp:
        example: "2026-01-19T15:37:27.514667373Z"
        type: string
    type: object
...
  response.PlaceDoc:
    properties:
      configuration_id:
        example: 1
        type: integer
      created_at:
        example: "2026-01-19T15:37:27.514Z"
        type: string
      id:
        example: 1
        type: integer
      is_active:
        example: true
        type: boolean
      label:
        example: PC-1
        type: string
      sort_order:
        example: 1
        type: integer
      updated_at:
        example: "2026-01-19T15:37:27.514Z"
        type: string
      zone_id:
        example: 2
        type: integer
    type: object
  response.PlaceListResponse:
    properties:
      places:
        items:
          $ref: '#/definitions/response.PlaceDoc'
        type: array
    type: object
  response.PlaceResponse:
    properties:
      place:
        $ref: '#/definitions/response.PlaceDoc'
    type: object
...
host: localhost
info:
  contact: {}
  description: Backend API for PlayGround
  title: PlayGround API
  version: "1.0"
paths:
...
  /api/v1/place:
    get:
      description: Returns all places
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PlaceListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: List places
      tags:
      - places
    post:
      consumes:
      - application/json
      description: Creates a place from JSON payload
      parameters:
      - description: Place payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/domain.CreatePlaceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PlaceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Create place
      tags:
      - places
  /api/v1/place/{id}:
    delete:
      description: Deletes place by id
      parameters:
      - description: Place ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DeletePlaceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Delete place
      tags:
      - places
    get:
      description: Returns place by id from path param
      parameters:
      - description: Place ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PlaceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get place by id
      tags:
      - places
    patch:
      consumes:
      - application/json
      description: Partially updates place fields by id
      parameters:
      - description: Place ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: Place patch payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/domain.PatchPlaceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.PlaceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Patch place
      tags:
      - places
...
swagger: "2.0"